move Hash model
