+++ fakeredis/model/__init__.py
@@ -2,6 +2,7 @@ from ._stream import XStream, StreamEntryKey, StreamGroup, StreamRangeTest
 from ._timeseries_model import TimeSeries, TimeSeriesRule, AGGREGATORS
 from ._topk import HeavyKeeper
 from ._zset import ZSet
+from ._hash import Hash
 
 __all__ = [
     "XStream",
@@ -13,4 +14,5 @@ __all__ = [
     "TimeSeriesRule",
     "AGGREGATORS",
     "HeavyKeeper",
+    "Hash",
 ]
